- Added new animation for the sidebar - Fixed z-index issues - Applied some media queries for about page
